statd: split code into multiple files (cosmetic)
Mainly in preparation for upcoming patches for ethtool.
