+++ src/statd/statd.c
@@ -22,8 +22,9 @@
 #include <srx/helpers.h>
 #include <srx/lyx.h>
 
-#define XPATH_MAX PATH_MAX
-#define XPATH_IFACE_BASE "/ietf-interfaces:interfaces"
+#include "shared.h"
+#include "iface-ip-link.h"
+#include "iface-ip-addr.h"
 
 #define SOCK_RMEM_SIZE 1000000  /* Arbitrary chosen, default = 212992 */
 #define NL_BUF_SIZE 4096 /* Arbitrary chosen */
@@ -106,654 +107,13 @@ static void sub_delete(struct ev_loop *loop, struct sub_head *subs, struct sub *
 	free(sub);
 }
 
-static json_t *_json_get_ip_output(const char *cmd)
-{
-	json_error_t j_err;
-	json_t *j_root;
-	FILE *proc;
-
-	proc = popenf("re", cmd);
-	if (!proc) {
-		ERROR("Error, running ip link command");
-		return NULL;
-	}
-
-	j_root = json_loadf(proc, 0, &j_err);
-	pclose(proc);
-	if (!j_root) {
-		ERROR("Error, parsing ip link JSON");
-		return NULL;
-	}
-
-	if (!json_is_array(j_root)) {
-		ERROR("Expected a JSON array from ip link");
-		json_decref(j_root);
-		return NULL;
-	}
-
-	return j_root;
-}
-
-static json_t *json_get_ip_link(const char *ifname)
+static json_t *json_get_ip_link(void)
 {
 	char cmd[512] = {}; /* Size is arbitrary */
 
-	if (ifname)
-		snprintf(cmd, sizeof(cmd), "ip -s -d -j link show dev %s 2>/dev/null", ifname);
-	else
-		snprintf(cmd, sizeof(cmd), "ip -s -d -j link show 2>/dev/null");
+	snprintf(cmd, sizeof(cmd), "ip -s -d -j link show 2>/dev/null");
 
-	return _json_get_ip_output(cmd);
-}
-
-static json_t *json_get_ip_addr(const char *ifname)
-{
-	char cmd[512] = {}; /* Size is arbitrary */
-
-	if (ifname)
-		snprintf(cmd, sizeof(cmd), "ip -j addr show dev %s 2>/dev/null", ifname);
-	else
-		snprintf(cmd, sizeof(cmd), "ip -j addr show 2>/dev/null");
-
-	return _json_get_ip_output(cmd);
-}
-
-static const char *get_yang_origin(const char *protocol)
-{
-	size_t i;
-	struct {
-		const char *kern;
-		const char *yang;
-
-	} map[] = {
-		{"kernel_ll",	"link-layer"},
-		{"static",	"static"},
-		{"dhcp",	"dhcp"},
-		{"random",	"random"},
-	};
-
-	for (i = 0; i < sizeof(map) / sizeof(map[0]); i++) {
-		if (strcmp(protocol, map[i].kern) != 0)
-			continue;
-
-		return map[i].yang;
-	}
-
-	return "other";
-}
-
-static const char *get_yang_operstate(const char *operstate)
-{
-	size_t i;
-	struct {
-		const char *kern;
-		const char *yang;
-
-	} map[] = {
-		{"DOWN",                "down"},
-		{"UP",                  "up"},
-		{"DORMANT",             "dormant"},
-		{"TESTING",             "testing"},
-		{"LOWERLAYERDOWN",      "lower-layer-down"},
-		{"NOTPRESENT",          "not-present"},
-	};
-
-	for (i = 0; i < sizeof(map) / sizeof(map[0]); i++) {
-		if (strcmp(operstate, map[i].kern) != 0)
-			continue;
-
-		return map[i].yang;
-	}
-
-	return "unknown";
-}
-
-static int ly_add_ip_link_stat(const struct ly_ctx *ctx, struct lyd_node **parent,
-			       char *xpath, json_t *j_iface)
-{
-	struct lyd_node *node = NULL;
-	char *stat_xpath;
-	json_t *j_stat;
-	json_t *j_val;
-	int err;
-
-	err = asprintf(&stat_xpath, "%s/statistics", xpath);
-	if (err == -1) {
-		ERROR("Error, creating statistics xpath");
-		return SR_ERR_SYS;
-	}
-
-	j_stat = json_object_get(j_iface, "stats64");
-	if (!j_stat) {
-		ERROR("Didn't find 'stats64' in JSON data");
-		goto err_out;
-	}
-
-	err = lyd_new_path(*parent, ctx, stat_xpath, NULL, 0, &node);
-	if (err) {
-		ERROR("Failed adding 'statistics' node, libyang error %d: %s",
-		      err, ly_errmsg(ctx));
-		goto err_out;
-	}
-
-	j_val = json_object_get(j_stat, "tx");
-	if (!j_val) {
-		ERROR("Didn't find 'tx' in JSON stats64 data");
-		goto err_out;
-	}
-
-	j_val = json_object_get(j_val, "bytes");
-	if (!j_val) {
-		ERROR("Didn't find 'bytes' in JSON stats64 tx data");
-		goto err_out;
-	}
-	if (!json_is_integer(j_val)) {
-		ERROR("Didn't get integer for 'bytes' in JSON stats64 tx data");
-		goto err_out;
-	}
-	err = lydx_new_path(ctx, &node, stat_xpath, "out-octets", "%lld",
-			    json_integer_value(j_val));
-	if (err) {
-		ERROR("Error, adding 'out-octets' to data tree, libyang error %d", err);
-		goto err_out;
-	}
-
-	j_val = json_object_get(j_stat, "rx");
-	if (!j_val) {
-		ERROR("Didn't find 'rx' in JSON stats64 data");
-		goto err_out;
-	}
-
-	j_val = json_object_get(j_val, "bytes");
-	if (!j_val) {
-		ERROR("Didn't find 'bytes' in JSON stats64 rx data");
-		goto err_out;
-	}
-	if (!json_is_integer(j_val)) {
-		ERROR("Didn't get integer for 'bytes' in JSON stats64 rx data");
-		goto err_out;
-	}
-	err = lydx_new_path(ctx, &node, stat_xpath, "in-octets", "%lld",
-			    json_integer_value(j_val));
-	if (err) {
-		ERROR("Error, adding 'in-octets' to data tree, libyang error %d", err);
-		goto err_out;
-	}
-
-	free(stat_xpath);
-
-	return SR_ERR_OK;
-
-err_out:
-	free(stat_xpath);
-
-	return SR_ERR_SYS;
-}
-
-static const char *get_yang_link_type(char *xpath, json_t *iface)
-{
-	json_t *j_type;
-	const char *type;
-
-	j_type = json_object_get(iface, "link_type");
-	if (!json_is_string(j_type)) {
-		ERROR("Expected a JSON string for 'link_type'");
-		/* This will throw a YANG / ly error */
-		return "";
-	}
-
-	type = json_string_value(j_type);
-
-	if (strcmp(type, "ether") == 0) {
-		const char *kind;
-		json_t *j_val;
-
-		j_val = json_object_get(iface, "linkinfo");
-		if (!j_val)
-			return "infix-if-type:ethernet";
-
-		j_val = json_object_get(j_val, "info_kind");
-		if (!j_val)
-			return "infix-if-type:ethernet";
-
-		if (!json_is_string(j_val)) {
-			ERROR("Expected a JSON string for 'info_kind'");
-			return "infix-if-type:other";
-		}
-		kind = json_string_value(j_val);
-
-		if (strcmp(kind, "veth") == 0)
-			return "infix-if-type:veth";
-		if (strcmp(kind, "vlan") == 0)
-			return "infix-if-type:vlan";
-		if (strcmp(kind, "bridge") == 0)
-			return "infix-if-type:bridge";
-
-		/**
-		 * We could return ethernetCsmacd here, but it might hide some
-		 * special type that we actually want to explicitly identify.
-		 */
-		ERROR("Unable to determine info_kind for \"%s\"", xpath);
-
-		return "infix-if-type:other";
-	}
-
-	if (strcmp(type, "loopback") == 0)
-		return "infix-if-type:loopback";
-
-	ERROR("Unable to determine infix-if-type for \"%s\"", xpath);
-
-	return "infix-if-type:other";
-}
-
-static int ly_add_ip_link_br(const struct ly_ctx *ctx, struct lyd_node **parent,
-			     char *xpath, json_t *j_iface)
-{
-	struct lyd_node *br_node = NULL;
-	char *br_xpath;
-	json_t *j_val;
-	int err;
-
-	j_val = json_object_get(j_iface, "master");
-	if (!j_val) {
-		/* Interface has no bridge */
-		return SR_ERR_OK;
-	}
-
-	if (!json_is_string(j_val)) {
-		ERROR("Expected a JSON string for bridge 'master'");
-		return SR_ERR_SYS;
-	}
-
-	err = asprintf(&br_xpath, "%s/infix-interfaces:bridge-port", xpath);
-	if (err == -1) {
-		ERROR("Error, creating bridge xpath");
-		return SR_ERR_SYS;
-	}
-
-	err = lyd_new_path(*parent, ctx, br_xpath, NULL, 0, &br_node);
-	if (err) {
-		ERROR("Failed adding 'bridge' node (%s), libyang error %d: %s",
-		      br_xpath, err, ly_errmsg(ctx));
-		free(br_xpath);
-		return SR_ERR_LY;
-	}
-
-	err = lydx_new_path(ctx, &br_node, br_xpath, "bridge",
-			"%s", json_string_value(j_val));
-	if (err) {
-		ERROR("Error, adding 'bridge' to data tree, libyang error %d", err);
-		free(br_xpath);
-		return SR_ERR_LY;
-	}
-	free(br_xpath);
-
-	return SR_ERR_OK;
-}
-
-static int iface_has_flag(json_t *j_iface, const char *flag)
-{
-	json_t *j_flags, *j_flag;
-	size_t i;
-
-	j_flags = json_object_get(j_iface, "flags");
-	if (!j_flags)
-		return 0;
-
-	if (!json_is_array(j_flags))
-		return 0;
-
-	json_array_foreach(j_flags, i, j_flag) {
-		if (strcmp(json_string_value(j_flag), flag) == 0)
-			return 1;
-	}
-
-	return 0;
-}
-
-static int ly_add_ip_link_data(const struct ly_ctx *ctx, struct lyd_node **parent,
-			       char *xpath, json_t *j_iface)
-{
-	const char *val;
-	const char *type;
-	json_t *j_val;
-	int err;
-
-	j_val = json_object_get(j_iface, "ifindex");
-	if (!json_is_integer(j_val)) {
-		ERROR("Expected a JSON integer for 'ifindex'");
-		return SR_ERR_SYS;
-	}
-
-	err = lydx_new_path(ctx, parent, xpath, "if-index", "%lld",
-			    json_integer_value(j_val));
-	if (err) {
-		ERROR("Error, adding 'if-index' to data tree, libyang error %d", err);
-		return SR_ERR_LY;
-	}
-
-	val = iface_has_flag(j_iface, "UP") ? "up" : "down"; /* 'testing' not supported */
-	err = lydx_new_path(ctx, parent, xpath, "admin-status", val);
-	if (err) {
-		ERROR("Error, adding 'admin-status' to data tree, libyang error %d", err);
-		return SR_ERR_LY;
-	}
-
-	j_val = json_object_get(j_iface, "operstate");
-	if (!json_is_string(j_val)) {
-		ERROR("Expected a JSON string for 'operstate'");
-		return SR_ERR_SYS;
-	}
-
-	val = get_yang_operstate(json_string_value(j_val));
-	err = lydx_new_path(ctx, parent, xpath, "oper-status", val);
-	if (err) {
-		ERROR("Error, adding 'oper-status' to data tree, libyang error %d", err);
-		return SR_ERR_LY;
-	}
-
-	err = ly_add_ip_link_stat(ctx, parent, xpath, j_iface);
-	if (err) {
-		ERROR("Error, adding 'stats64' to data tree");
-		return err;
-	}
-
-	type = get_yang_link_type(xpath, j_iface);
-	err = lydx_new_path(ctx, parent, xpath, "type", type);
-	if (err) {
-		ERROR("Error, adding 'type' to data tree, libyang error %d", err);
-		return SR_ERR_LY;
-	}
-
-	j_val = json_object_get(j_iface, "address");
-	if (!json_is_string(j_val)) {
-		ERROR("Expected a JSON string for 'address'");
-		return SR_ERR_SYS;
-	}
-
-	err = lydx_new_path(ctx, parent, xpath, "phys-address", json_string_value(j_val));
-	if (err) {
-		ERROR("Error, adding 'phys-address' to data tree, libyang error %d", err);
-		return SR_ERR_LY;
-	}
-
-	err = ly_add_ip_link_br(ctx, parent, xpath, j_iface);
-	if (err) {
-		ERROR("Error, adding bridge to data tree, libyang error %d", err);
-		return err;
-	}
-
-	j_val = json_object_get(j_iface, "link");
-	if (j_val) {
-		if (!json_is_string(j_val)) {
-			ERROR("Expected a JSON string for 'link'");
-			return SR_ERR_SYS;
-		}
-
-		err = lydx_new_path(ctx, parent, xpath,
-				    "ietf-if-extensions:parent-interface", "%s",
-				    json_string_value(j_val));
-		if (err) {
-			ERROR("Error, adding 'link' to data tree, libyang error %d", err);
-			return SR_ERR_LY;
-		}
-	}
-
-	return SR_ERR_OK;
-}
-
-static int ly_add_ip_mtu(const struct ly_ctx *ctx, struct lyd_node *node,
-			   char *xpath, json_t *j_iface, char *ifname)
-{
-	json_t *j_val;
-	int err;
-
-	/**
-	 * TODO: Not sure how to handle loopback MTU (65536) which is
-	 * out of bounds for both YANG and uint16_t. For now, we skip it.
-	 */
-	if (strcmp(ifname, "lo") == 0)
-		return SR_ERR_OK;
-
-	j_val = json_object_get(j_iface, "mtu");
-	if (!json_is_integer(j_val)) {
-		ERROR("Expected a JSON integer for 'mtu'");
-		return SR_ERR_SYS;
-	}
-
-	err = lydx_new_path(ctx, &node, xpath, "mtu", "%lld",
-			    json_integer_value(j_val));
-	if (err) {
-		ERROR("Error, adding 'mtu' to data tree, libyang error %d", err);
-		return SR_ERR_LY;
-	}
-
-	return SR_ERR_OK;
-}
-
-static int ly_add_ip_addr_origin(const struct ly_ctx *ctx, struct lyd_node *addr_node,
-				      char *addr_xpath, json_t *j_addr)
-{
-	const char *origin;
-	json_t *j_val;
-	int err;
-
-	/* It's okey not to have an origin */
-	j_val = json_object_get(j_addr, "protocol");
-	if (!j_val)
-		return SR_ERR_OK;
-
-	if (!json_is_string(j_val)) {
-		ERROR("Expected a JSON string for ip 'protocol'");
-		return SR_ERR_SYS;
-	}
-
-	origin = get_yang_origin(json_string_value(j_val));
-
-	/**
-	 * kernel_ll/link-layer only has a link-layer origin if its address is
-	 * based on the link layer address (addrgenmode eui64).
-	 */
-	if (strcmp(origin, "link-layer") == 0) {
-		j_val = json_object_get(j_addr, "stable-privacy");
-		if (j_val && json_is_boolean(j_val) && json_boolean_value(j_val))
-			origin = "random";
-	}
-
-	err = lydx_new_path(ctx, &addr_node, addr_xpath, "origin", "%s", origin);
-	if (err) {
-		ERROR("Error, adding ip 'origin' to data tree, libyang error %d", err);
-		return SR_ERR_LY;
-	}
-
-	return SR_ERR_OK;
-}
-
-static int ly_add_ip_addr_info(const struct ly_ctx *ctx, struct lyd_node *node,
-			       char *xpath, json_t *j_addr)
-{
-	struct lyd_node *addr_node = NULL;
-	char *addr_xpath;
-	const char *ip;
-	json_t *j_val;
-	int err;
-
-	j_val = json_object_get(j_addr, "local");
-	if (!json_is_string(j_val)) {
-		ERROR("Expected a JSON string for ip 'local'");
-		return SR_ERR_SYS;
-	}
-	ip = json_string_value(j_val);
-
-	err = asprintf(&addr_xpath, "%s/address[ip='%s']", xpath, ip);
-	if (err == -1) {
-		ERROR("Error, creating address xpath");
-		return SR_ERR_SYS;
-	}
-
-	err = lyd_new_path(node, ctx, addr_xpath, NULL, 0, &addr_node);
-	if (err) {
-		ERROR("Failed adding ip 'address' node (%s), libyang error %d: %s",
-		      addr_xpath, err, ly_errmsg(ctx));
-		free(addr_xpath);
-		return SR_ERR_LY;
-	}
-
-	j_val = json_object_get(j_addr, "prefixlen");
-	if (!json_is_integer(j_val)) {
-		ERROR("Expected a JSON integer for ip 'prefixlen'");
-		free(addr_xpath);
-		return SR_ERR_SYS;
-	}
-
-	err = lydx_new_path(ctx, &addr_node, addr_xpath, "prefix-length",
-			    "%lld", json_integer_value(j_val));
-	if (err) {
-		ERROR("Error, adding ip 'prefix-length' to data tree, libyang error %d", err);
-		free(addr_xpath);
-		return SR_ERR_LY;
-	}
-
-	err = ly_add_ip_addr_origin(ctx, addr_node, addr_xpath, j_addr);
-
-	free(addr_xpath);
-
-	return err;
-}
-
-static int ly_add_ip_data(const struct ly_ctx *ctx, struct lyd_node **parent,
-			  char *xpath, int version, json_t *j_iface, char *ifname)
-{
-	struct lyd_node *node = NULL;
-	json_t *j_val;
-	json_t *j_addr;
-	size_t index;
-	int err;
-
-	err = lyd_new_path(*parent, ctx, xpath, NULL, 0, &node);
-	if (err) {
-		ERROR("Failed adding 'ip' node (%s), libyang error %d: %s",
-		      xpath, err, ly_errmsg(ctx));
-		return SR_ERR_LY;
-	}
-
-	err = ly_add_ip_mtu(ctx, node, xpath, j_iface, ifname);
-	if (err) {
-		ERROR("Error, adding ip MTU");
-		return err;
-	}
-
-	j_val = json_object_get(j_iface, "addr_info");
-	if (!json_is_array(j_val)) {
-		ERROR("Expected a JSON array for 'addr_info'");
-		return SR_ERR_SYS;
-	}
-
-	json_array_foreach(j_val, index, j_addr) {
-		json_t *j_family;
-
-		j_family = json_object_get(j_addr, "family");
-		if (!json_is_string(j_family)) {
-			ERROR("Expected a JSON string for ip 'family'");
-			return SR_ERR_SYS;
-		}
-
-		if (version == 4 && (strcmp(json_string_value(j_family), "inet") == 0)) {
-			err = ly_add_ip_addr_info(ctx, node, xpath, j_addr);
-			if (err) {
-				ERROR("Error, adding  address");
-				return err;
-			}
-		} else if (version == 6 && (strcmp(json_string_value(j_family), "inet6") == 0)) {
-			err = ly_add_ip_addr_info(ctx, node, xpath, j_addr);
-			if (err) {
-				ERROR("Error, adding  address");
-				return err;
-			}
-		}
-	}
-
-	return SR_ERR_OK;
-}
-
-static int ly_add_ip_link(const struct ly_ctx *ctx, struct lyd_node **parent, char *ifname)
-{
-	char xpath[XPATH_MAX] = {};
-	json_t *j_root;
-	json_t *j_iface;
-	int err;
-
-	j_root = json_get_ip_link(ifname);
-	if (!j_root) {
-		ERROR("Error, parsing ip-link JSON");
-		return SR_ERR_SYS;
-	}
-	if (json_array_size(j_root) != 1) {
-		ERROR("Error, expected JSON array of single iface");
-		json_decref(j_root);
-		return SR_ERR_SYS;
-	}
-
-	j_iface = json_array_get(j_root, 0);
-
-	snprintf(xpath, sizeof(xpath), "%s/interface[name='%s']",
-		 XPATH_IFACE_BASE, ifname);
-
-	err = ly_add_ip_link_data(ctx, parent, xpath, j_iface);
-	if (err) {
-		ERROR("Error, adding ip-link info for %s", ifname);
-		json_decref(j_root);
-		return err;
-	}
-
-	json_decref(j_root);
-
-	return SR_ERR_OK;
-}
-
-static int ly_add_ip_addr(const struct ly_ctx *ctx, struct lyd_node **parent, char *ifname)
-{
-	char xpath[XPATH_MAX] = {};
-	json_t *j_root;
-	json_t *j_iface;
-	int err;
-
-	j_root = json_get_ip_addr(ifname);
-	if (!j_root) {
-		ERROR("Error, parsing ip-addr JSON");
-		return SR_ERR_SYS;
-	}
-	if (json_array_size(j_root) != 1) {
-		ERROR("Error, expected JSON array of single iface");
-		json_decref(j_root);
-		return SR_ERR_SYS;
-	}
-
-	j_iface = json_array_get(j_root, 0);
-
-	snprintf(xpath, sizeof(xpath), "%s/interface[name='%s']/ietf-ip:ipv4",
-		 XPATH_IFACE_BASE, ifname);
-
-	err = ly_add_ip_data(ctx, parent, xpath, 4, j_iface, ifname);
-	if (err) {
-		ERROR("Error, adding ipv4 addr info for %s", ifname);
-		json_decref(j_root);
-		return err;
-	}
-
-	snprintf(xpath, sizeof(xpath), "%s/interface[name='%s']/ietf-ip:ipv6",
-		 XPATH_IFACE_BASE, ifname);
-
-	err = ly_add_ip_data(ctx, parent, xpath, 6, j_iface, ifname);
-	if (err) {
-		ERROR("Error, adding ipv6 addr info for %s", ifname);
-		json_decref(j_root);
-		return err;
-	}
-
-	json_decref(j_root);
-
-	return SR_ERR_OK;
+	return json_get_output(cmd);
 }
 
 static int sr_ifaces_cb(sr_session_ctx_t *session, uint32_t, const char *path,
@@ -945,7 +305,7 @@ static int sub_to_ifaces(struct statd *statd)
 	json_t *j_root;
 	size_t i;
 
-	j_root = json_get_ip_link(NULL);
+	j_root = json_get_ip_link();
 	if (!j_root) {
 		ERROR("Error, parsing ip-link JSON");
 		return SR_ERR_SYS;
